:art: Improve stracture
