+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-core/src/test/java/com/alibaba/cloud/ai/graph/serializer/AppendableValueRW.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.cloud.ai.graph.state;
+package com.alibaba.cloud.ai.graph.serializer;
 
 import java.io.Externalizable;
 import java.io.IOException;
