+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-core/src/test/java/com/alibaba/cloud/ai/graph/serializer/ObjectInputWithMapper.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.cloud.ai.graph.serializer.std;
+package com.alibaba.cloud.ai.graph.serializer;
 
-import com.alibaba.cloud.ai.graph.serializer.Serializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
