+++ spring-ai-alibaba-studio/src/main/java/com/alibaba/cloud/ai/service/GraphService.java
@@ -16,7 +16,7 @@
 package com.alibaba.cloud.ai.service;
 
 import com.alibaba.cloud.ai.graph.GraphInitData;
-import com.alibaba.cloud.ai.graph.GraphStateException;
+import com.alibaba.cloud.ai.graph.exception.GraphStateException;
 import com.alibaba.cloud.ai.graph.StateGraph;
 import com.alibaba.cloud.ai.param.GraphStreamParam;
 import org.springframework.http.codec.ServerSentEvent;
