+++ spring-ai-alibaba-studio/src/main/java/com/alibaba/cloud/ai/service/impl/GraphServiceImpl.java
@@ -18,7 +18,7 @@ package com.alibaba.cloud.ai.service.impl;
 import com.alibaba.cloud.ai.graph.CompileConfig;
 import com.alibaba.cloud.ai.graph.CompiledGraph;
 import com.alibaba.cloud.ai.graph.GraphRepresentation;
-import com.alibaba.cloud.ai.graph.GraphStateException;
+import com.alibaba.cloud.ai.graph.exception.GraphStateException;
 import com.alibaba.cloud.ai.graph.NodeOutput;
 import com.alibaba.cloud.ai.graph.PersistentConfig;
 import com.alibaba.cloud.ai.graph.RunnableConfig;
