+++ spring-ai-alibaba-studio/src/test/java/com/alibaba/cloud/ai/graph/GraphStudioApplication.java
@@ -19,7 +19,7 @@
 // import com.alibaba.cloud.ai.graph.action.EdgeAction;
 // import com.alibaba.cloud.ai.graph.prebuilt.MessagesState;
 // import com.alibaba.cloud.ai.graph.prebuilt.MessagesStateGraph;
-// import com.alibaba.cloud.ai.graph.state.AgentState;
+// import com.alibaba.cloud.ai.graph.serializer.AgentState;
 // import org.checkerframework.checker.units.qual.C;
 //
 // import org.springframework.boot.SpringApplication;
